+++ lib/grassroots_bluetooth_layer.dart
@@ -4,7 +4,9 @@ export 'src/grassroots_bluetooth_layer.dart' show GrassrootsBluetooth;
 export 'src/generated/grassroots_bluetooth_layer.g.dart'
     show
         BleAdapterState,
+        BleAdvertiseFailure,
         BleAdvertisement,
+        BleAdvertisingState,
         BleLinkInfo,
         BlePath,
         BlePathState,